replacing python-oauth2 with rauth
It would seem that rauth has become one of the most used and active
OAuth implementations for Python while python-oauth2 has not updated
in at least a year and is considered abandonware by many. Rauth is
based on simple and clean design principles and makes use of the
excellent Python Requests library. So de facto, rauth seems to be
the preferred choice. It then seems reasonable to update this list as
such.