Create short URL for the OAuth Security Workshop
At https://sec.uni-stuttgart.de/events/osw2019, we will soon publish more information and the registration for the next OAuth Security Workshop, which in 2019 will be held at the University of Stuttgart, Germany. It would be great to create a short URL to the conference page, and update it for the workshops in the next years.