Add Omejdn to Open Source Authorization Servers
About Omejdn:
"Omejdn" is bavarian for "Log In". The Server written in Ruby tries to support both OAuth 2.0/2.1 and OpenID Connect while following the Security BCP, along with many current features such as Server Metadata, JWT Access Tokens, Pushed Authorization Requests, ...
It currently does not support Dynamic Client Registration and Management.

It tries to maintain a minimal code base and to be easily extensible, making it suitable as a base for research projects.