Add link to Local Implementation in metadata document
This project was created out of necessity while working on CIMD integration in MCP Rust SDK.There is no open-source implementation of the CIMD protocol available for testing. This minimalist implementation provides only the necessary
features required by the CIMD specification (MUST / MUST NOT rules), along with a complete client application, making it easy for developers to understand and test the protocol locally.